fix: make performance acceptable for complex networks
This involved adding an explicit matched component (zero reflection)
for ports, eliminating edges that have zero weight, recalculating cycles
on every edge repopulation, and making it possible to exchange a and b
so that the s-parameters looking into a *port* (not into the match element)
can be calculated. That is, added the ability to calculate the s-params
looking out of an element as well as into an element. Because of implicit
T-networks, the difference can be important (ex: branchline coupler).

Also made all the components Copy-able, to simplify the ergonomics of
testing. Added a script to clean up DOT notation for graph debugging.
Changed a whole bunch of function visibilities to prevent caching
and ordering mistakes.